requirements.txt
requests
beautifulsoup4
python-dotenv
streamlit
watchdog
streamlit_autorefresh